requirements.txt
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
streamlit==1.28.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
chromadb==0.4.10
huggingface-hub==0.19.0